fix error for mac

+++ zsh/.bashrc
@@ -350,8 +350,7 @@ alias journalctl='SYSTEMD_LESS=FRSMK journalctl'
 ########### oh-my-linux ######################
 #if grep -Ei "Arch Linux" /etc/issue > /dev/null || grep -Ei "Arch Linux" /etc/*-release > /dev/null; then
 if { [[ -f /etc/issue ]] && grep -Ei "Arch Linux" /etc/issue > /dev/null; } || \
-   { ls /etc/*-release >/dev/null 2>&1 && grep -Ei "Arch Linux" /etc/*-release > /dev/null; }; then
-#if [[ "$(uname -a)" == *Linux* ]]; then
+   { [[ -f /etc/os-release ]] && grep -Ei "Arch Linux" /etc/os-release > /dev/null; }; then
 	alias reboot='sudo systemctl reboot'
 	alias poweroff='sudo systemctl poweroff'
 	alias suspend='sudo systemctl suspend'
